Finished overhauling ECDSA code
The ECDSA code now combines addresses, public keys and private keys into
a single object called "AddressData" (though I believe the name must be
changed).  The benefit of this class is better-encapsulation of related
methods, and removing redundant data.  It also supports serialization
and unserialization of key data, useful for wallet files.

The serialization process will store only whatever data is available in
the AddressData object.  If private keys are not available, it will only
store addresses and public keys.  If pubkeys are not available, then
only addresses.

This enables having a single file format for all types of accounts.
We may be simply tracking addresses from multiple computers/devices,
but don't want our private keys to be distributed all over the place.
We can serialize a version of our address data that doesn't include
private keys, and then that file can be more widely distributed for
convenience.