Some profiling done.  Commit before branching to test/profile non-fancy header-reading solutions.
For my AMD Phenom II X4 840, within a Windows XP VM:
    BHM::getHash():  1.5 microseconds per exec (~650/1300 KHash/s)
    BHM::importHeadersFromBlkFile():     1.18s
    BHM::importHeadersFromHeaderFile():  0.75s
    BHM::organizeChain():                0.68s
It appears that the fancier header-reading solution is a bit faster.  If the block-DATA reading is slow, this might be worth investigating more.  Still need to try a more-straightforward solution of just keeping a linked-list of headers (yes, I *should've* done that first...)