Got self-updating hash dicts, CLI options...
If the flag is set at execution, will read in all the hashes from
blk0001.dat, and save the last read location, so that next time it is
run it only reads the new data.

Still some issues with finding PkScripts, and the hash dictionary gets
populated but we never match anything in it.