Working, fast, entire-blockchain-in-RAM C++
The C++ code can now read the entire blockchain (as of block 135k) and find/construct the longest chain in less than 15s.  This is a critical capability for the python code, which could take on the order of minutes to do it.  Haven't tested tx-scanning or wallet management yet.  But all the pieces I need are implemented, just need to do it.

A few important things to be done:
-- Reorganize the code so it's not one giant, complicate source file
-- Will eventually implement a technique that isn't FullRAM
-- Merge into the master branch
-- Make intuitive interface for SWIG -- so that the C++ code can serve as a "BlockChain service" for python
-- Actually implement SWIG bindings...

Going to reorganize the code a bit, and then merge into the