Fixed a nasty SegFault in TxIn/BinaryData
Also added some functions for helping trace Tx's down the chain.
Previously, it was very difficult to find the source of a TxIn, because
it's not always apparent from its script, so you need the BDM's help to
go find the TxOut.  There was a nasty BDM bug involving negative values
of unsigned integers that occurred here... but it's good now and finding
the sender of transactions is a lot easier now!