Proof of concept getting C++ data to PyQt!
The GUI actually loads the blockchain and displays the top block and the
first Tx.  No response to user events, and default behaviors of widgets
are terrible.  But anything is possible, now!

I am probably going to have to create a combined-C++-python wrapper,
because the two libraries don't play nicely right now (it would be nice
if I could get the BinaryData typemaps working in SWIG...).