Reorg blockdata-reading code, still need debugging
I created a unified interface for adding blockdata to the memory pool.
This was a little complicated because I needed to make it clear which
data was being copied, and which data was permanently in memory, and
not to get them mixed up when copying pointers.  This also means that
if we can add data from the blk0001.dat, we can add new raw blocks,
since it's the same code.