Super-annoying bug when you forget to open a file in binary mode...
My unit test was failing because I forgot to open the file in binary, and the first ctrl-D char it hit, it stopped reading and filled in the rest of the space with zero-bytes.  Of course, no warning or seg-fault...

Also updated the unit test so that the tx's that are supposed to be in both chains are actually the _same_ txs.  I had been accidentally re-signing these transactions, which led to a different hash, and thus treated as a new tx instead of a recycled one.