Script parsing fixed 100% -- need to test eval now
Had some issues with the OP_PUSHDATAX codes, which caused it to throw
an error if the data was at the end of the script.  This bug was in
both the C++ and Python and it's been fixed in both.

So far, I've found lots of super-interesting scripts, but haven't
actually tried evaluating any of them.  That will be my next task,
though it looks like I need to implement the OP_IF/_ELSE/_ENDIF/etc
first...