FINALLY SelectCoins->TxDP->Sign pipeline working!
I was able to plug in my private key, a recipient address, and value to
send that address, and my code did the rest.  It even did so through my
TxDP class, which means this could've easily been done from an offline
wallet (instead of signing the TxDP right here, just serialize it to a
USB key, sign it offline, and then bring it back to broadcast it.

Still plenty of work left to do, but this is a big milestone!