Addr encryption/[un]lock/changingKey works great!
Finished unit-testing all the code for encrypting, unencrypting,
locking, unlocking, and changing encryption key of a particular
PyBtcAddress object.  Still need to test reading in pre-encrypted key
data, and then chained-key generation (deterministic addresses).  After
that, I'll be ready for PyBtcWallet (which should be a hell of a lot
easier with all this implemented).