Finished deterministic addr gen tests! It works!
Tested the most complicated part of the deterministic logic, which
is the case where you have a private-key-bearing address, but you
don't have the encryption key when you want to generate new addrs.
You can generate the public key and be confident you can find the
private key next time you unlock the address/wallet, but it can be
difficult to juggle the parameters that allow me to pretend I've
got the private keys without actually having them yet.