PyBtcAddr [un]serialize works with all encrypt ops
The unit test demonstrates just about every feature of the PyBtcAddress
class, with all encryption, deterministic generation, and serialization
and unserialization that can be used for storing wallet files.  This
unit test is exhaustive, testing all the complicated features.  As such,
I can finally move onto the PyBtcWallet class which should be
dramatically simpler with all this logic worked out in the addr class.