Corruption-resistent wallet files, almost finished!
Maintaining a synchronous backup wallet, and touching/deleting
flag-files to make sure we know when, and which, files are potentially
corrupted.  We effectively make wallet updates "atomic" and can
guarantee that we always have a GOOD version of the wallet.  Just be
sure that you do your file operations BEFORE giving users access to the
data (such as generating new addresses:  if you fail at writing the
data to file, there is no partial recovery:  you restore from backup
that doesn't yet contain the new address).