Updated initial scan to include relevant wallets
The C++ code for the initial blockchain scan now accepts a pointer
to a wallet, and will collect the tiny subset of all tx it crosses
that are relevant to that tx.  That means, we only have to scan
a few DOZEN tx after indexing the blockchain, instead of a few MILLION.
Since we are using mmap() now instead of full-RAM, this is critically
important!