Fixed recv-warning dialog, and impl qrc-resources!
This is version 0.61 and has some small-yet-substantial improvements!

(1) Moved all images into a resource file so that it is all accessible
without the img/ directory, only import qrc_img_resources.py.  The UI
has been updated to use the resources file (fixed a problem in Linux if
you run Armory from a different directory)

(2) Fixed the receive-coins-with-watching-only-wallet warning window.
If you canceled from the warning window it would pop up a blank window
that needed to be closed.  It doesn't do that anymore!

(3) Crypto++ is now statically compiled into the _CppBlockUtils.so on
Linux and OSX (it always has been statically compiled in Windows).  This
solves a variety of nuisances when compiling, linking and running Armory
on *nix systems.