Start testing UI upgrades for RAM-reduction
RAM-reduction has occurred at the C++ level, with the blockchain
intelligently rescanning only if necessary.  But such a rescan will
take a long time now, meaning the user needs to be prompted and confirm
that's what they want to do.

Therefore, new UI elements have to be developed to handle confirmations,
and reorder operations that were assumed to be fast but aren't anymore.