MMAP'd Armory works!  Version (0.70-beta)-alpha
I have finally got the full MMAP solution implemented for Linux & OSX
and got the GUI updated to accommodate making the user wait if a rescan
is required to continue (or cancel out if they don't want to wait).

I just noticed an issue that I don't think is new:  wallet migration
is stupid slow for some reason (encrypted only?).  I will look at that
soon.

I will request *nix users to help test this version, and I will work on
the Windows-specific code branch while they are doing that.  It should
be only a matter of branching the BinaryDataMMAP class, and then all the
pointer stuff will work out the same.  I can't wait to find out it's
not...