Fix exception when double-clicking non-standard transactions in the ledger.
Transaction f3676c60a3498bf3f1d630f0b5350f54e4ea7781599dcab0432ea92c09aa52a1 is
one transaction where this occurs, and it appears to be because of the strange
zero-amount txout that accompanies every p2pool generation transaction.

See the backtrace below.  Basically, the problem is that
TxOutScriptExtractAddr160(), despite its name, doesn't necessarily return an
Addr160, but can return other things, depending on the type of the
transaction.  In the case of a TXOUT_SCRIPT_UNKNOWN transaction, it returns
'<Non-standard TxOut script>', which breaks base58_to_binary.

I see this as a flaw in the design of the TxOutScriptExtractAddr160 function,
and this patch does not fix that flaw, but it fixes the exception.

Traceback (most recent call last):
  File "ArmoryQt.py", line 1866, in dblClickLedger
    DlgDispTxInfo( pytx, self.walletMap[wltID], self, self, txtime=txtime).exec_()
  File "/home/dwon/sw/bitcoin/BitcoinArmory/qtdialogs.py", line 6600, in __init__
    self.txOutModel = TxOutDispModel(self.pytx,  self.main, idxGray=indicesMakeGray)
  File "/home/dwon/sw/bitcoin/BitcoinArmory/armorymodels.py", line 549, in __init__
    self.wltIDList.append(main.getWalletForAddr160(recip160))
  File "ArmoryQt.py", line 1024, in getWalletForAddr160
    if wlt.hasAddr(addr160):
  File "/home/dwon/sw/bitcoin/BitcoinArmory/armoryengine.py", line 5820, in hasAddr
    return self.addrMap.has_key(addrStr_to_hash160(addrData))
  File "/home/dwon/sw/bitcoin/BitcoinArmory/armoryengine.py", line 646, in addrStr_to_hash160
    return base58_to_binary(binStr)[1:-4]
  File "/home/dwon/sw/bitcoin/BitcoinArmory/armoryengine.py", line 619, in base58_to_binary
    n += BASE58CHARS.index(ch)
ValueError: substring not found