Removed -march flag from crypto++ proj Makefile.
It turns out that march=native produces code that is only guaranteed to
work on the CPU on which it was compiled.  Removing the option entirely
means gcc will compile with a more-generic instruction set that is
likely to work on the most common CPUs.  I don't even know why
march=native even worked at all!