Started testing full app.  1/1000 tx missed!?!
Out of hundreds of transactions over all my wallets,  It appears that
there is only a single transaction that was missed in the new code,
compared to the old code!  Exactly one.  Going to have to investigate...

And I just realized I never tested "duplicate" transactions in the
multi-map.  I guess that's next...