LICENSE using QFile causing hard crashes! Made .py
I guess I don't know how to use QFile with qrc_resources.  It
worked on my 10.04 64-bit machine, but causes bad_alloc/coredump
on all my VMs.  Instead I just turned LICENSE into LICENSE.py and
have the function licenseText() return the full text of it.
It is necessary to do something like this, because I don't want
to bloat the other source files with the license text (and it
needs to be in its own files), but I can't have it as a floating
file, because I'll never know where to find it when I need it
(if it's an importable python file, it will always be found...)