Initial multi-thread testing going extremely well
One thing I still need to figure out is, if I request a scan or rescan,
the BDM doesn't have time to change self.blkMode before the calling
thread starts a while loop based on that variable.  It will see blkMode
is "Full" and then not wait at all.  Need to figure out a clean way for
the calling thread to block until the BDM thread has time to process the
request