Commented the hell out of new threading code
Started updating the code outside of BlockDataManagerThread to access
BDM intelligently.  It's not always obvious how to do it:  should I
block?  Or queue the request asynchronously and check back later?
Which conditions should I use in which cases?  If someone else was
using this code, what would they expect the default behavior to be?

Lot's left to do, but at least the aynchronous BDM works!  It just
might require some care to make sure we don't accidentally block
the main GUI thread when we intended to kick off processing in the
background...