SWIG bug seg-fault workaround in address book
There appears to be a SWIG bug that causes seg faults sometimes when
iterating over C++ vector<...> and list<...> containers when using
the "-threads" option when generating swig files:

http://sourceforge.net/tracker/?func=detail&atid=101645&aid=3403085&group_id=1645

I'm not sure why it doesn't trigger all the time in the new code, but
it triggered 100% of the time on the address book when trying to
iterate over a vector<RegisteredTx> object.  I put in a workaround,
so that I no longer have "for regTx in abe.getTxList(): ...", and the
fix does work.  But there is lots of other code that iterates that way,
and I need to do lots of testing to out any other seg-faults...