Multi-threaded programming is a bitch.
The wallet recovery/restore scan code is getting silly.  In order
for the BDM to orchestrate the recovery scan, it needs to call
PyBtcWallet methods directly, which are NORMALLY invoked outside
the BDM and use the BDM-queue.  I've had to branch the code to
allow PyBtcWallet methods to call BDM methods directly in the
event that the "useDirectBDM" flag is set (which will be set by
the BDM before it makes any PyBtcWallet calls).

The code is becoming more complicated than I had hoped.  I also
believe that there are better ways to do this... I'm thrashing
around a bit trying to figure the best way to do it -- only
comforted by the fact that I know this is necessarily a complicated
process, so I know that messy is not strictly incorrect...