Finally got wallet recovery! Burned by error checking.
It turns out I had disabled a bunch of PyBtcWallet methods when the
BDM is busy... but when those methods are called by the BDM itself,
the BDM is always busy!  Adding an extra conditional to check whether
we were called by the BDM avoids this.

Recovery rescans appear to work now.  If any addresses are detected
in the last half of the pre-generated key pool, we will extend the
keypool and rescan again.  When we extend the keypool but find no
extra transactions, that's when we're done.