Updated blkXXXX.dat detect for Bitcoin-Qt 0.8+
This was actually quite a pain in the ass.  It wasn't just updating to
use 0.8+ path convention, but accommodating users who might use either
one, or even some users that switch to 0.8 (thus creating the files),
then switch back to pre-0.8.

Therefore, the code checks a few different locations, and observes the
latest timestamp for the highest blkxxxx.dat file.  Then that is passed
into the BDM, along with how many digits to use and the start index of
the first file.  Had to add quite a few threads to the C++ to make this
happen.