Timers appear to work.  Now start profiling!
In fact, I timed the operation of starting and stopping timers, to see
if it will burden any methods, even ones that are very very fast.  The
result was 100,000 timer start/stops in 0.12sec on my system.  So it
can do about one million per sec.

Also, I was originally concerned about the efficiency of coin2str, since
I had modified it somewhat inefficiently in order to guarantee I got
what I wanted out of it.  It does about 1,000,000 coin2str ops in 5 sec.
Not fantastic, but it's not what is slowing down my GUI...