Started a new proxy class for LevelDB/multimap ops
I'me hoping that the proxy class will be allow me to define identical
operations for both the multimap<> (RAM) solution and the LevelDB (HDD)
solution.  A simple preprocessor would let you switch between them at
compile time.

It's definitely more complicated than just updating the code directly to
use LevelDB... But it may be worth it in the long run.