I can't believe the SDM actually kinda works!
(Linux) It successfully creates the datadir, finds bitcoind, starts
the process, then monitors it and returns reasonable getSDMState()
values.  It fully sync'd testnet in a few minutes and looked like
it was ready to go!

There are still some issues needing to be fixed -- such as sometimes
it doesn't shutdown the process when Armory exits.  Probably a simple
Popen arg for that.  Also, will need to figure out file permissions
on Windows so that I can safely create a bitcoin.conf file for new
users that don't have one yet -- but it must be set with strict
permissions or it could open up a security hole!

The proof-of-concept is done.  Just gotta grind through the rest...