Taking picobit's suggestion, created guardian.py
It's very simple, but seems very powerful.  The idea is that when Armory
starts bitcoind in a subprocess, there's no guarantee bitcoind shuts
down when Armory hard-crashes (killed, segfault, etc).   However, once
it is started, Armory can grab its own pid, and the pid of bitcoind and
pass it to a separate "guardian" process.  If Armory doesn't shut down
bitcoind, it won't shut down the guardian either ... which will shutdown
bitcoind and then itself if it detects that Armory closed.