FINALLY got a working guardian.exe
In windows, subprocess.Popen actually seems to open a cmd.exe process, which spawns the process you actually wanted to run.  When you get the PID of the Popen object, you only get the cmd.exe process, and killing it does not kill the process you started.

The guardian executable now gets the PID of cmd.exe and searches the process list looking for a process that has the PID as its parent.  It then adds that to the list of processes to be killed.  So far it works!