Correct balance for first 5 blocks of reorg test!
Still got a couple minor issues to resolve, but this is an excellent
milestone: the database is properly tracking non-reorg balances.  Now
will have to test the reorg part.

For some reason, it's running the last block twice, and also the var
appliedToDB is not getting written when it's applied.

Next up:  reorg/undo testing.