Looks like about 3 hours with 850MB of RAM to build DB
More specifically:

    1 min to read all the headers and organize the chain
   40 min to write all the raw blocks to disk
    3-4 hours to update the DB with spentness info and SSHs
   850 MB used (corresponds to 128 MB UPDATE_BYTES_THRESH (cache)

That's usable, and can probably even be parallelized with the initial
blockchain download so that it won't even be noticeable.  Increase times
by about 25% and it uses less than 400 MB to do the full indexing.

However, it segfaults immediately after finishing:  I'm going to have
to get creative with the unit tests to catch that one...