All MSVS projects upgraded to 2012, including LevelDB
This includes LevelDB+Snappy compression library, albeit the
versions that were available when the leveldbwin project was
created (2011).  The versions of the source will be updated
as soon as we confirm that the compiled projects do what they
are supposed to do.  For now, *anything* that compiles LevelDB
is a huge step forward, as LevelDB is notoriously difficult
to integrate into a Windows build system