Improved notification info, still not great
Unfortunately, the notification system doesn't pass
around raw scripts so I couldn't use the new methods
created, getDisplayStringForScript().  Nonethelss,
displayed info on notifications needed to be fixed.

It is a lot better, though I still occasionally get
instances where it shows me the change part of the
transaction in-addtion-to/instead-of the part we care
about.  This will need some work.