Added code coverage to the gtest project
This is for linux, using gcc. Make sure you install gcov and lcov,
then build and run the gtest project.  When it's done, run
computecov.sh in that directory, which will build an HTML index
of the covered code.  It will then automatically pop up firefox
with the coverage data loaded.