Updated qtdialogs.py 2 clearify defau. URL handler
Without specifying the type of URL handler, i.e. (Bitcoin URL), the users gets the impression that Armory is requesting to handle any urls including canonical ones like http. Although the description below does specify the type of url, the size and font weight of the header can obfuscate it and thus contribute to unnecessary friction.