Fix armoryd logic error.
armoryd was maintaining a dictionary w/ wallet/ID & lockbox/ID combos
and a set that had just the IDs. A wallet or lockbox, when built by
armoryd, was only added to the dictionary, while some code would check
the set. The set has been removed, with all affected code checking the
dictionary keys.