fixes ARMORY-264
Changed the fancy arrow (→) to a more standard dash-gt (->) in order for the tr() function to not give the error. If we really want to use the unicode character, we need the default to be UTF-8 or something to make it work. For now, this will not barf when you try to go to File -> Settings.