webshop 2.0
Now can process refunds. There is a refund link which allows the user to
request a refund. The unsigned transaction is generated automatically and
entered into the refunds directory. The auto-generated tx uses only the inputs
from the payment, thus eliminates double-spending risk. The transaction can
then be signed by someone later to actually put the refund on the network.

There is still a small security hole in that if the attacker knows the address
the payment was sent to, they can ask for a refund. The code assumes that the
actual signer of the refund will verify that the customer is actually the one
asking for the refund in some other manner (email is probably best).