Added automatic versioning tool:
There is a script called update_version.py in the root directory which
will change the file armoryengine/ArmoryBuild.py. The only thing in this
file is the constant BTCARMORY_BUILD.

The purpose of this variable is to contain the git hash of the code that
it's running. When developing, this can be found at startup on armoryd or
ArmoryQT or in the UI at Help -> About Armory as the last part of the
version number shown.