Autotools: ArmoryMac shared object build
Adds a rule to the Makefile (if we are targeting OS X) to build
ArmoryMac.so. For now you have to manually run make objc and copy the
resulting shared object to the root of the Armory directory, but the
goal is to make this part of another rule that creates a .app.