Autotools: Qt 5 cross build script for OS X
The build-qt5.sh script creates a Qt 5 build for OS X in
osxbuild/install-qt5/. The script currently works when building on
Linux. Support for building on OS X is to be added. The resulting OS X
build will be used to embed the Qt frameworks in the Armory.app that is
built by the autotools-based build system, but currently nothing is done
with Qt 5 once it is built.