Release: Just use first file for reproducible verification
This causes just the first file in the list of output files in the
assert file to be used. Practically speaking we never have multiple
output files per build. And other parts of the code already made the
assumption that there is one output file per build. So this didn't work
as expected anyway. This can always be revisited if we need to.