Deb: Deb package fixes
Pipe yes into dh_make to make script more automatic.

Make downloading of dependencies more robust. Use apt-get download to
grab dependencies in a way where hardcoded URLs won't break (because now
there are no hardcoded URLs). Uses apt config file to run apt-get in a
self-contained way such that it won't mess with existing user
configuration. Currently hardcodes the package names and doesn't have a
method of determining the dependency tree (dependencies of the
dependencies, etc.), so if there is an update to the list of Armory
dependencies, the script has to be modified to reflect that.