Fix Armory minimization issue on macOS
When running Armory in mainnet mode and setting Armory to minimize to the tray when clicking the close button, the Armory icon in the macOS dock can only close Armory, and not restore it. Fix this problem and remove the need for the workaround (use the Armory icon in the menu bar to restore).